Change input of `Constraint.for_dataset()`
Previously this method only got a `Dataset`-type argument. However,
for path resolution (for example), this information is insufficient,
because DataLad needs to know whether the type of the original argument.

For this reason `DatasetParameter` already existed, and is now also used
for this purpose.

This also simplified the type-annotation -- no need for custom typevars.

It also made sense to move the `DatasetParameter` class to `base.py`,
because every single constraint implementation has to support it (or
use the default implementation from `base.py`.
